rename interaction circle to interaction
