+++ src/entities/adventurer/prefab.ts
@@ -4,7 +4,7 @@ import { AdventurerComponent } from "../../components/adventurer-component";
 import { AttachmentComponent } from "../../components/attachment-component";
 import { HurtboxComponent } from "../../components/hurtbox-component";
 import { BoundsComponent } from "../../components/bounds-component";
-import { InteractionCircleComponent } from "../../components/interaction-circle-component";
+import { InteractionComponent } from "../../components/interaction-component";
 import { PhiniteStateMachineComponent } from "../../components/phinite-state-machine-component";
 import { ShootsArrowsComponent } from "../../components/shoots-arrows-component";
 import { InvulnerabilityComponent } from "../../components/invulnerability-component";
@@ -45,7 +45,7 @@ export const adventurerPrefab: Phecs.Prefab = {
       }
     },
     {
-      component: InteractionCircleComponent,
+      component: InteractionComponent,
       data: {
         interactionRadius: 30,
         interactionDebug: false,
