+++ src/entities/door/prefab.ts
@@ -1,7 +1,7 @@
 import { SpriteComponent } from "../../components/sprite-component";
 import { DoorComponent } from "../../components/door-component";
 import { SpriteIndicatorComponent } from "../../components/sprite-indicator-component";
-import { InteractionCircleComponent } from "../../components/interaction-circle-component";
+import { InteractionComponent } from "../../components/interaction-component";
 import { AttachmentComponent } from "../../components/attachment-component";
 
 export const doorPrefab: Phecs.Prefab = {
@@ -23,7 +23,7 @@ export const doorPrefab: Phecs.Prefab = {
       }
     },
     {
-      component: InteractionCircleComponent,
+      component: InteractionComponent,
       data: {
         interactionRadius: 40,
       }
