+++ src/entities/npc/prefab.ts
@@ -1,6 +1,6 @@
 import { SpriteComponent } from "../../components/sprite-component";
 import { PhiniteStateMachineComponent } from "../../components/phinite-state-machine-component";
-import { InteractionCircleComponent } from "../../components/interaction-circle-component";
+import { InteractionComponent } from "../../components/interaction-component";
 import { SpriteIndicatorComponent } from "../../components/sprite-indicator-component";
 import { ConversationComponent } from "../../components/conversation-component";
 import { NpcComponent } from "../../components/npc-component";
@@ -17,7 +17,7 @@ export const npcPrefab: Phecs.Prefab = {
       }
     },
     {
-      component: InteractionCircleComponent,
+      component: InteractionComponent,
       data: {
         interactionRadius: 70,
         interactionDebug: false,
