+++ src/entities/sheep/prefab.ts
@@ -2,7 +2,7 @@ import { SpriteComponent } from "../../components/sprite-component";
 import { PhysicsBodyComponent } from "../../components/physics-body-component";
 import { PhiniteStateMachineComponent } from "../../components/phinite-state-machine-component";
 import { ZoneBoundaryComponent } from "../../components/zone-boundary-component";
-import { InteractionCircleComponent } from "../../components/interaction-circle-component";
+import { InteractionComponent } from "../../components/interaction-component";
 import { ConversationComponent } from "../../components/conversation-component";
 import { SayComponent } from "../../components/say-component";
 import { AttachmentComponent } from "../../components/attachment-component";
@@ -38,7 +38,7 @@ export const sheepPrefab: Phecs.Prefab = {
       }
     },
     {
-      component: InteractionCircleComponent,
+      component: InteractionComponent,
       data: {
         interactionRadius: 30,
         interactionDebug: false,
